Calculate crowdsource status in a single query
This avoids pulling crowdsourced status to ruby.

The query is built in a way that we could change it to update every
store status in a single query by just removing the statuses.id = #{id}
part.

updated_time and updated_at are using DateTime.now instead of NOW()
because the latter was causing tests to fail. This is probably related
to the conditions in services/calculate_status.rb.